refactor: :recycle: Consolidated socket connection logic into SocketManager
- Deleted SocketService class and the corresponding socket-service.ts file
- Moved connection management, event tracking, and emit/on methods directly into SocketManagerBase
- Added a new method waitForConnection to SocketManagerBase
- Updated ApiKeyValidator and chat-socket-listener.ts to reference SocketManager instead of the removed SocketService type
